Extract convert rules from md-converter
