+++ src/lib/convert-rules.spec.js
@@ -1,0 +1,7 @@
+const { link, image, code, codespan } = require('./convert-rules');
+
+describe('stub', () => {
+    it('stub', () => {
+
+    });
+});
