--- src/lib/md-converter.spec.js
@@ -1,7 +1,0 @@
-const { md2html } = require('./md-converter');
-
-describe('stub', () => {
-    it('stub', () => {
-
-    });
-});
